Fix warning on msvc

+++ libr/util/base36.c
@@ -1,11 +1,11 @@
-/* radare - LGPL - Copyright 2023 - pancake */
+/* radare - LGPL - Copyright 2023-2024 - pancake */
 
 #include <r_util.h>
 
 static const char cb36[] = "0123456789abcdefghijklmnopqrstuvwxyz";
 
 R_API void b36_fromnum(char *s, ut64 n) {
-	const int amount = n? (log ((double)n) / log ((double)36)): 0;
+	const int amount = n? (int)(log ((double)n) / log ((double)36)): 0;
 	char *p = s;
 	*p++ = '0';
 	*p++ = '_';
